requirements.txt
azure-identity==1.22.0
numpy==2.2.5
openai==1.77.0
rich==13.6.0
PyYAML==6.0.2
